storage: Rename indexed_db.rs to indexeddb.rs for naming consistency
Renames components/script/indexed_db.rs to indexeddb.rs to maintain consistent
naming across Servo. The codebase already uses the indexeddb form in most
locations, so this change aligns indexedb.rs with the rest of the code.

Testing: Unit tests continue to pass.
