+++ components/script/dom/indexeddb/idbkeyrange.rs
@@ -14,7 +14,7 @@ use crate::dom::bindings::error::{Error, Fallible};
 use crate::dom::bindings::import::module::SafeJSContext;
 use crate::dom::bindings::reflector::{Reflector, reflect_dom_object};
 use crate::dom::globalscope::GlobalScope;
-use crate::indexed_db::{convert_value_to_key, key_type_to_jsval};
+use crate::indexeddb::{convert_value_to_key, key_type_to_jsval};
 
 #[dom_struct]
 pub struct IDBKeyRange {
